Views Refactoring
- Introduce ViewProvider trait
- Move all show_* in navigation.rs into separate file that implements ViewProvider
- Move some helpers out from navigation.rs into providers/mod.rs
