+++ src/view/mod.rs
@@ -3,7 +3,10 @@ mod log_view;
 mod navigation;
 mod process_view;
 mod processes_view;
+mod provider;
+pub mod providers;
 mod query_result_view;
+mod registry;
 mod summary_view;
 mod text_log_view;
 
@@ -11,8 +14,10 @@ pub use navigation::Navigation;
 pub use process_view::ProcessView;
 pub use processes_view::ProcessesView;
 pub use processes_view::Type as ProcessesType;
+pub use provider::ViewProvider;
 pub use query_result_view::QueryResultView;
 pub use query_result_view::Row as QueryResultRow;
+pub use registry::ViewRegistry;
 pub use summary_view::SummaryView;
 
 pub use ext_table_view::ExtTableView;
