+++ src/view/providers/replicated_fetches.rs
@@ -1,0 +1,44 @@
+use crate::{
+    interpreter::{ContextArc, options::ChDigViews},
+    view::ViewProvider,
+};
+use cursive::Cursive;
+use std::collections::HashMap;
+
+pub struct ReplicatedFetchesViewProvider;
+
+impl ViewProvider for ReplicatedFetchesViewProvider {
+    fn name(&self) -> &'static str {
+        "Fetches"
+    }
+
+    fn view_type(&self) -> ChDigViews {
+        ChDigViews::ReplicatedFetches
+    }
+
+    fn show(&self, siv: &mut Cursive, context: ContextArc) {
+        let mut columns = vec![
+            "database",
+            "table",
+            "result_part_name part",
+            "elapsed",
+            "progress",
+            "total_size_bytes_compressed size",
+            "bytes_read_compressed bytes",
+        ];
+
+        // TODO: on_submit show last related log messages
+        super::show_query_result_view(
+            siv,
+            context,
+            "replicated_fetches",
+            None,
+            None,
+            "elapsed",
+            &mut columns,
+            3,
+            Some(super::query_result_show_row),
+            &HashMap::new(),
+        );
+    }
+}
